+++ cs276-pa3/src/edu/stanford/cs276/doc/BodyTermFreqExtractor.java
@@ -13,16 +13,16 @@ import java.util.Map;
  */
 public class BodyTermFreqExtractor extends TermFreqExtractor {
     @Override
-    public Map<String, Double> extractFrom(Document d, Query q) {
+    public Map<String, Integer> extractFrom(Document d, Query q) {
         Map<String, List<Integer>> bodyHits = d.getBodyHits();
         if (bodyHits.size() == 0) {
             return Collections.emptyMap();
         }
 
-        Map<String, Double> bodyTF = new HashMap<>();
+        Map<String, Integer> bodyTF = new HashMap<>();
 
         for (Map.Entry<String, List<Integer>> et : bodyHits.entrySet()) {
-            bodyTF.put(et.getKey(), (double) et.getValue().size());
+            bodyTF.put(et.getKey(), et.getValue().size());
         }
 
         return bodyTF;
